New version of GPIBLogger.py
Fixing a problem where the logger would actually run slower than shown by timestamps in the log file. This could happen because of delays on the GPIB interface. With the fix, the time stamps are always taken from the host computer, so if the logger runs slow, say every 1.2 seconds instead of every second, the time stamps are now showing the true interval of 1.2 instead of 1 as before.